create controller for getCreatorDetails function
